refactor: more hole_punch to room_creation.rs in ctrl_chnl
